+++ packages/core/src/lib/implementation/runner.ts
@@ -1,14 +1,12 @@
 import path from 'node:path';
-import type {
-  OnProgress,
-  RunnerConfig,
-  RunnerFunction,
-} from '@code-pushup/models';
+import type { RunnerConfig, RunnerFunction } from '@code-pushup/models';
 import {
   calcDuration,
   executeProcess,
+  isVerbose,
   readJsonFile,
   removeDirectoryIfExists,
+  ui,
 } from '@code-pushup/utils';
 
 export type RunnerResult = {
@@ -19,7 +17,6 @@ export type RunnerResult = {
 
 export async function executeRunnerConfig(
   cfg: RunnerConfig,
-  onProgress?: OnProgress,
 ): Promise<RunnerResult> {
   const { args, command, outputFile, outputTransform } = cfg;
 
@@ -27,7 +24,14 @@ export async function executeRunnerConfig(
   const { duration, date } = await executeProcess({
     command,
     args,
-    observer: { onStdout: onProgress },
+    observer: {
+      onStdout: stdout => {
+        if (isVerbose()) {
+          ui().logger.log(stdout);
+        }
+      },
+      onStderr: stderr => ui().logger.error(stderr),
+    },
   });
 
   // read process output from file system and parse it
@@ -48,13 +52,12 @@ export async function executeRunnerConfig(
 
 export async function executeRunnerFunction(
   runner: RunnerFunction,
-  onProgress?: OnProgress,
 ): Promise<RunnerResult> {
   const date = new Date().toISOString();
   const start = performance.now();
 
   // execute plugin runner
-  const audits = await runner(onProgress);
+  const audits = await runner();
 
   // create runner result
   return {