+++ testing/test-setup/src/lib/reset.mocks.ts
@@ -3,6 +3,10 @@ import { beforeEach, vi } from 'vitest';
 
 beforeEach(() => {
   vi.clearAllMocks();
+
   vi.unstubAllEnvs();
+  // set the verbose to false as default for all tests as local env could be set to true
+  vi.stubEnv('CP_VERBOSE', 'false');
+
   vol.reset();
 });