+++ testing/test-utils/src/lib/utils/git.ts
@@ -9,12 +9,15 @@ import type {
 import { vi } from 'vitest';
 
 export type GitConfig = { name: string; email: string };
-
 export async function initGitRepo(
   simpleGit: SimpleGitFactory,
-  opt: { baseDir: string; config?: GitConfig },
+  opt: {
+    baseDir: string;
+    config?: GitConfig;
+    baseBranch?: string;
+  },
 ): Promise<SimpleGit> {
-  const { baseDir, config } = opt;
+  const { baseDir, config, baseBranch } = opt;
   const { email = 'john.doe@example.com', name = 'John Doe' } = config ?? {};
   await mkdir(baseDir, { recursive: true });
   const git = simpleGit(baseDir);
@@ -23,7 +26,7 @@ export async function initGitRepo(
   await git.addConfig('user.email', email);
   await git.addConfig('commit.gpgSign', 'false');
   await git.addConfig('tag.gpgSign', 'false');
-  await git.branch(['-M', 'main']);
+  await git.branch(['-M', baseBranch ?? 'main']);
   return git;
 }